+++ src/ui.rs
@@ -3,7 +3,7 @@ use itertools::{Itertools, repeat_n};
 use ratatui::{
     layout::Constraint::{Fill, Length},
     prelude::*,
-    widgets::{Block, Borders, List, ListItem, Paragraph, Scrollbar, ScrollbarState},
+    widgets::{Block, Borders, List, ListItem, Paragraph, Scrollbar, ScrollbarState, Wrap},
 };
 
 use crate::{
@@ -71,7 +71,7 @@ fn render_topics(frame: &mut Frame, area: Rect, model: &Model, filter: Option<&F
     }))
     .block(
         Block::new()
-            .title(Line::raw("Topics").italic().gray())
+            .title(Line::raw("Topics").italic().dark_gray())
             .borders(Borders::TOP),
     );
 
@@ -100,37 +100,64 @@ fn render_details(
     jq: &Jaqqer,
     scroll: u16,
 ) {
-    let [header, pane, footer] = Layout::vertical([
-        Length(1),
+    let message = model.message(topic).unwrap_or_default();
+    let error = model.error(topic);
+
+    let [header, pane, warning, footer] = Layout::vertical([
+        Length(2),
         Fill(0),
+        Length(error.map(|e| e.lines().count() + 1).unwrap_or_default() as u16),
         Length(if jq.is_dormant() { 0 } else { 6 }),
     ])
     .areas(area);
 
     let [header, indicator] = Layout::horizontal([Fill(0), Length(2)]).areas(header);
-
     let [details, scroller] = Layout::horizontal([Fill(0), Length(1)]).areas(pane);
 
     // Top header with topic name
-    frame.render_widget(Paragraph::new(topic).bold().centered(), header);
-    frame.render_widget(connection_status(model).right_aligned(), indicator);
+    frame.render_widget(
+        Paragraph::new(topic).bold().centered().block(
+            Block::new()
+                .title(Line::raw("Message").italic().dark_gray())
+                .borders(Borders::BOTTOM),
+        ),
+        header,
+    );
+    frame.render_widget(
+        connection_status(model)
+            .right_aligned()
+            .block(Block::new().borders(Borders::BOTTOM).gray()),
+        indicator,
+    );
 
     let mut style = Style::new();
     if model.is_copy() {
         style = style.reversed();
     }
 
-    let message = model.message(topic).unwrap_or_default();
     frame.render_widget(
-        Paragraph::new(model.highlight(&message, details, scroll).style(style))
-            .scroll((scroll, 0))
-            .block(
-                Block::new()
-                    .title(Line::raw("Message").italic().gray())
-                    .borders(Borders::TOP),
-            ),
+        Paragraph::new(
+            if error.is_none() {
+                model.highlight(&message, details, scroll)
+            } else {
+                message.clone().into()
+            }
+            .style(style),
+        )
+        .scroll((scroll, 0))
+        .wrap(Wrap { trim: false }),
         details,
     );
+    if let Some(error) = error {
+        frame.render_widget(
+            Paragraph::new(Text::from(error).dark_gray()).block(
+                Block::new()
+                    .title(Line::raw("Warning").italic().yellow())
+                    .borders(Borders::TOP),
+            ),
+            warning,
+        )
+    }
     frame.render_stateful_widget(
         Scrollbar::new(ratatui::widgets::ScrollbarOrientation::VerticalRight)
             .begin_symbol(None)
@@ -140,7 +167,6 @@ fn render_details(
             .position(scroll as usize),
     );
 
-    // let jq = jq.render(frame, footer).block();
     let filter = match jq {
         Jaqqer::Dormant => Default::default(),
         Jaqqer::Prompt {