+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Show connection status in header
 * Automatically reconnect if broker restarts
 * `-q/--quit` do immediately disconnect on connection loss
+* Allow non-JSON messages but show warning
 
 ## 1.0.0